Add files via upload
These are all projects I've done throughout high school for class or practice. Although initially trained in Java, I have since translated the skills learned in these projects to C++ and, more importantly, Python.

I apologize for the lack of comments for most of the code. Hopefully, the projects are understandable enough based on their titles, except for those with Hebrew or transliterated Hebrew titles that were made for my classes.